move composite classes in their own package
