+++ app/src/main/java/com/nbusy/app/data/Chat.java
@@ -1,6 +1,7 @@
 package com.nbusy.app.data;
 
 import com.google.common.collect.ImmutableSet;
+import com.nbusy.app.data.composite.ChatAndNewMessages;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -58,23 +59,6 @@ public final class Chat {
         }
 
         return new ChatAndNewMessages(upsertMessages(newMsgs), ImmutableSet.copyOf(newMsgs));
-    }
-
-    public final class ChatAndNewMessages {
-        public final Chat chat;
-        public final ImmutableSet<Message> messages;
-
-        public ChatAndNewMessages(Chat chat, ImmutableSet<Message> messages) {
-            if (chat == null) {
-                throw new IllegalArgumentException("chat cannot be null");
-            }
-            if (messages == null || messages.size() == 0) {
-                throw new IllegalArgumentException("messages cannot be null or empty");
-            }
-
-            this.chat = chat;
-            this.messages = messages;
-        }
     }
 
     Chat upsertMessages(List<Message> msgs) {
